+++ src/schick/seg082.h
@@ -5,7 +5,7 @@ namespace M302de {
 extern struct struct_chest g_dng07_specialchests[2];
 
 /* 0x20 */
-signed short DNG07_handler(void);
+signed int DNG07_handler(void);
 /* 0x25 */
 void DNG09_statues(const signed int, const signed int);
