+++ src/schick/seg090.cpp
@@ -37,7 +37,7 @@ struct struct_chest g_dng12_specialchests[3] = {
 signed int g_dng12_obstacle_tries = 0; // ds:0x9d43 SHOULD BE IN GAME STATE
 signed int g_dng12_watertrap_bak = 0;  // ds:0x9d45 SHOULD BE IN GAME STATE
 
-signed short DNG12_handler(void)
+signed int DNG12_handler(void)
 {
 	signed int target_pos;
 	signed int i;
