+++ src/schick/seg090.h
@@ -9,7 +9,7 @@ void DNG_clear_corridor(int8_t*);
 /* 0x25 */
 void DNG12_chest00_loot(struct struct_chest*);
 /* 0x2a */
-signed short DNG12_handler(void);
+signed int DNG12_handler(void);
 /* 0x2f */
 void DNG12_fight_intro(const signed int);
